# openapi.yaml
openapi: 3.0.3
info:
  title: Excalidraw FULL API
  description: API for the visual workspace platform
  version: 1.0.0
servers:
  - url: http://localhost:3000
paths:
...
  /api/teams/{teamID}/invites:
    get:
      operationId: listTeamInvites
      security: [ { cookieAuth: [] } ]
      parameters: [ { name: teamID, in: path, required: true, schema: { type: string } } ]
      responses:
        '200':
          description: Invites
          content:
            application/json:
              schema: { type: array, items: { $ref: '#/components/schemas/TeamInvite' } }
    post:
      operationId: createTeamInvite
      security: [ { cookieAuth: [] } ]
      parameters: [ { name: teamID, in: path, required: true, schema: { type: string } } ]
      requestBody:
        required: true
        content:
          application/json:
            schema: { $ref: '#/components/schemas/CreateInviteRequest' }
      responses:
        '201':
          description: Created
          content:
            application/json:
              schema:
                type: object
                properties:
                  invite: { $ref: '#/components/schemas/TeamInvite' }
                  token: { type: string }
  /api/invites/accept:
    post:
      operationId: acceptInvite
      security: [ { cookieAuth: [] } ]
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required: [token]
              properties:
                token: { type: string }
      responses:
        '200':
          description: Membership created
          content:
            application/json:
              schema: { $ref: '#/components/schemas/TeamMember' }
...
  /api/drawings/{drawingID}/share-links:
    get:
      operationId: listShareLinks
      security: [ { cookieAuth: [] } ]
      parameters: [ { name: drawingID, in: path, required: true, schema: { type: string } } ]
      responses:
        '200':
          description: Links
          content:
            application/json:
              schema: { type: array, items: { $ref: '#/components/schemas/ShareLink' } }
    post:
      operationId: createShareLink
      security: [ { cookieAuth: [] } ]
      parameters: [ { name: drawingID, in: path, required: true, schema: { type: string } } ]
      requestBody:
        required: true
        content:
          application/json:
            schema: { $ref: '#/components/schemas/CreateShareLinkRequest' }
      responses:
        '201':
          description: Created
          content:
            application/json:
              schema:
                type: object
                properties:
                  share_link: { $ref: '#/components/schemas/ShareLink' }
                  token: { type: string }
...
  /api/drawings/{drawingID}/embeds:
    get:
      operationId: listEmbeds
      security: [ { cookieAuth: [] } ]
      parameters: [ { name: drawingID, in: path, required: true, schema: { type: string } } ]
      responses:
        '200':
          description: Embeds
          content:
            application/json:
              schema: { type: array, items: { $ref: '#/components/schemas/Embed' } }
    post:
      operationId: createEmbed
      security: [ { cookieAuth: [] } ]
      parameters: [ { name: drawingID, in: path, required: true, schema: { type: string } } ]
      requestBody:
        required: true
        content:
          application/json:
            schema: { $ref: '#/components/schemas/CreateEmbedRequest' }
      responses:
        '201':
          description: Created
          content:
            application/json:
              schema: { $ref: '#/components/schemas/Embed' }
...
components:
  securitySchemes:
    cookieAuth:
      type: apiKey
      in: cookie
      name: excalidraw_session
  schemas:
...
    TeamMember:
      type: object
      required: [id, team_id, user_id, role]
      properties:
        id: { type: string }
        team_id: { type: string }
        user_id: { type: string }
        role: { type: string, enum: [owner, admin, editor, viewer] }
        created_at: { type: string, format: date-time }
    TeamInvite:
      type: object
      required: [id, team_id, email, role, invited_by]
      properties:
        id: { type: string }
        team_id: { type: string }
        email: { type: string }
        role: { type: string }
        invited_by: { type: string }
        expires_at: { type: string, format: date-time }
        created_at: { type: string, format: date-time }
...
    ShareLink:
      type: object
      required: [id, resource_type, resource_id, access_level, created_by]
      properties:
        id: { type: string }
        resource_type: { type: string }
        resource_id: { type: string }
        access_level: { type: string, enum: [view, comment, edit] }
        password_hash: { type: string, nullable: true }
        expires_at: { type: string, format: date-time, nullable: true }
        created_by: { type: string }
        created_at: { type: string, format: date-time }
...
    Embed:
      type: object
      required: [id, drawing_id, url, type]
      properties:
        id: { type: string }
        drawing_id: { type: string }
        url: { type: string }
        type: { type: string }
        title: { type: string, nullable: true }
        thumbnail_url: { type: string, nullable: true }
        created_by: { type: string }
        created_at: { type: string, format: date-time }
...
    CreateInviteRequest:
      type: object
      required: [email, role]
      properties:
        email: { type: string, format: email }
        role: { type: string, enum: [admin, editor, viewer] }
...
    CreateShareLinkRequest:
      type: object
      required: [access_level]
      properties:
        access_level: { type: string, enum: [view, comment, edit] }
        password: { type: string, nullable: true }
        expires_at: { type: string, format: date-time, nullable: true }
...
    CreateEmbedRequest:
      type: object
      required: [url, type]
      properties:
        url: { type: string }
        type: { type: string }
        title: { type: string, nullable: true }
        thumbnail_url: { type: string, nullable: true }
...